+++ src/renderer/src/components/automations/AutomationListTableHeader.tsx
@@ -5,34 +5,85 @@ import {
   LIST_TABLE_HEADER_CLASS,
   LIST_TABLE_STICKY_HEADER_CELL_CLASS
 } from '@/lib/list-table-layout'
+import { AutomationListSortHeader } from './AutomationListSortHeader'
+import type { AutomationListSort, AutomationListSortField } from './automation-list-view'
 
-export function AutomationListTableHeader(): React.JSX.Element {
-  const labels = [
-    ['auto.components.automations.AutomationsPage.tableName', 'Name'],
-    ['auto.components.automations.AutomationDetail.18763ded26', 'Schedule'],
-    ['auto.components.automations.AutomationsPage.tableProject', 'Project'],
-    ['auto.components.automations.AutomationsPage.tableHost', 'Host'],
-    ['auto.components.automations.AutomationDetail.578ff46987', 'Next run'],
-    ['auto.components.automations.AutomationsPage.tableLastRun', 'Last run'],
-    ['auto.components.automations.AutomationsPage.tableStatus', 'Status'],
-    ['auto.components.automations.AutomationDetail.2df8970cd5', 'Agent']
-  ] as const
+type HeaderColumn = {
+  key: string
+  fallback: string
+  /** Absent for columns the list cannot order by. */
+  sortField?: AutomationListSortField
+}
+
+const COLUMNS: readonly HeaderColumn[] = [
+  {
+    key: 'auto.components.automations.AutomationsPage.tableName',
+    fallback: 'Name',
+    sortField: 'name'
+  },
+  {
+    key: 'auto.components.automations.AutomationDetail.18763ded26',
+    fallback: 'Schedule'
+  },
+  {
+    key: 'auto.components.automations.AutomationsPage.tableProject',
+    fallback: 'Project'
+  },
+  {
+    key: 'auto.components.automations.AutomationsPage.tableHost',
+    fallback: 'Host'
+  },
+  {
+    key: 'auto.components.automations.AutomationDetail.578ff46987',
+    fallback: 'Next run'
+  },
+  {
+    key: 'auto.components.automations.AutomationsPage.tableLastRun',
+    fallback: 'Last run',
+    sortField: 'lastRun'
+  },
+  {
+    key: 'auto.components.automations.AutomationsPage.tableStatus',
+    fallback: 'Status'
+  },
+  {
+    key: 'auto.components.automations.AutomationDetail.2df8970cd5',
+    fallback: 'Agent'
+  }
+]
+
+export function AutomationListTableHeader({
+  sort = null,
+  onSort
+}: {
+  sort?: AutomationListSort | null
+  onSort?: (field: AutomationListSortField) => void
+} = {}): React.JSX.Element {
   return (
     <div className={`${AUTOMATIONS_TABLE_GRID_CLASS} ${LIST_TABLE_HEADER_CLASS}`}>
-      {labels.map(([key, fallback], index) => (
-        <span
-          key={key}
-          className={
-            index === 0
-              ? LIST_TABLE_STICKY_HEADER_CELL_CLASS
-              : index === labels.length - 1
-                ? 'text-center'
-                : undefined
-          }
-        >
-          {translate(key, fallback)}
-        </span>
-      ))}
+      {COLUMNS.map((column, index) => {
+        const label = translate(column.key, column.fallback)
+        const className =
+          index === 0
+            ? LIST_TABLE_STICKY_HEADER_CELL_CLASS
+            : index === COLUMNS.length - 1
+              ? 'text-center'
+              : undefined
+        return (
+          <span key={column.key} className={className}>
+            {column.sortField && onSort ? (
+              <AutomationListSortHeader
+                field={column.sortField}
+                label={label}
+                sort={sort}
+                onSort={onSort}
+              />
+            ) : (
+              label
+            )}
+          </span>
+        )
+      })}
       <span className="sr-only">
         {translate('auto.components.automations.AutomationsPage.tableActions', 'Actions')}
       </span>
