+++ src/renderer/src/components/automations/automation-list-view-sort.test.ts
@@ -5,32 +5,34 @@ import {
   type AutomationListSort,
   type AutomationListViewItem
 } from './automation-list-view'
+import { unscopedAutomationListRows } from './automation-list-row-identity'
 import { makeAutomation } from './automations-page-fixtures'
-
-const locale = vi.hoisted(() => ({ value: 'en' }))
-vi.mock('@/i18n/i18n', () => ({ getIntlLocale: () => locale.value }))
 
 afterEach(() => {
   vi.restoreAllMocks()
-  locale.value = 'en'
 })
 
-function rows(count = 512): AutomationListViewItem[] {
+function items(count = 512): AutomationListViewItem[] {
   const names = ['Alpha', 'álpha', 'Ångström', 'Zebra', 'Örebro', 'I', 'ı', 'İ', 'job 10', 'job 2']
   return buildAutomationListViewItems({
-    automations: Array.from({ length: count }, (_, index) =>
-      makeAutomation({ id: `job-${index}`, name: names[(index * 7) % names.length] })
+    rows: unscopedAutomationListRows(
+      Array.from({ length: count }, (_, index) =>
+        makeAutomation({
+          id: `job-${index}`,
+          name: names[(index * 7) % names.length]
+        })
+      )
     ),
-    externalEntries: [],
-    runs: []
+    externalEntries: []
   })
 }
 
-function previousOrder(items: AutomationListViewItem[], sort: AutomationListSort) {
+/** The pre-collator comparator, resolving options on every comparison. */
+function previousOrder(list: AutomationListViewItem[], sort: AutomationListSort, locale: string) {
   function compare(left: AutomationListViewItem, right: AutomationListViewItem) {
     const value =
       sort.field === 'name'
-        ? left.name.localeCompare(right.name, locale.value, { sensitivity: 'base' })
+        ? left.name.localeCompare(right.name, locale, { sensitivity: 'base' })
         : (left.lastRunAt ?? 0) - (right.lastRunAt ?? 0)
     return value !== 0
       ? sort.direction === 'asc'
@@ -38,37 +40,35 @@ function previousOrder(items: AutomationListViewItem[], sort: AutomationListSort
         : -value
       : left.id.localeCompare(right.id)
   }
-  return [...items].sort(compare)
+  return [...list].sort(compare)
 }
 
 describe('automation list collation', () => {
   it.each(['en', 'sv', 'tr', 'ja'])(
     'preserves %s ordering, tie-breaks and input identity',
-    (language) => {
-      locale.value = language
-      const items = rows()
-      const original = [...items]
+    (locale) => {
+      const list = items()
+      const original = [...list]
       for (const direction of ['asc', 'desc'] as const) {
         const sort = { field: 'name', direction } as const
-        const expected = previousOrder(items, sort)
-        const result = sortAutomationListViewItems(items, sort)
+        const expected = previousOrder(list, sort, locale)
+        const result = sortAutomationListViewItems(list, sort, locale)
         expect(result).toEqual(expected)
         expect(result.every((row, index) => row === expected[index])).toBe(true)
       }
-      expect(items).toEqual(original)
+      expect(list).toEqual(original)
     }
   )
 
-  it('resolves collation once per name sort and responds to locale changes', () => {
-    const items = rows()
+  it('resolves collation once per name sort and follows the locale it is given', () => {
+    const list = items()
     const OriginalCollator = Intl.Collator
     const construct = vi.spyOn(Intl, 'Collator').mockImplementation(function (locales, options) {
       return new OriginalCollator(locales, options)
     })
     const compare = vi.spyOn(String.prototype, 'localeCompare')
-    sortAutomationListViewItems(items, { field: 'name', direction: 'asc' })
-    locale.value = 'sv'
-    sortAutomationListViewItems(items, { field: 'name', direction: 'desc' })
+    sortAutomationListViewItems(list, { field: 'name', direction: 'asc' }, 'en')
+    sortAutomationListViewItems(list, { field: 'name', direction: 'desc' }, 'sv')
     expect(construct.mock.calls).toEqual([
       ['en', { sensitivity: 'base' }],
       ['sv', { sensitivity: 'base' }]
@@ -76,16 +76,39 @@ describe('automation list collation', () => {
     expect(compare.mock.calls.filter((args) => args.length >= 3)).toHaveLength(0)
   })
 
+  it('orders by row key, not the bare automation ID, so hosts cannot collapse', () => {
+    const duplicate = makeAutomation({ id: 'shared', name: 'Same' })
+    const list = buildAutomationListViewItems({
+      rows: [
+        {
+          key: 'row|host-b|shared',
+          automation: duplicate,
+          hostLabel: 'b',
+          usageSummary: null
+        },
+        {
+          key: 'row|host-a|shared',
+          automation: duplicate,
+          hostLabel: 'a',
+          usageSummary: null
+        }
+      ],
+      externalEntries: []
+    })
+    const sorted = sortAutomationListViewItems(list, { field: 'name', direction: 'asc' }, 'en')
+    expect(sorted.map((item) => item.id)).toEqual(['row|host-a|shared', 'row|host-b|shared'])
+  })
+
   it('does not construct collation for unsorted, time-sorted or trivial lists', () => {
-    const items = rows()
+    const list = items()
     const construct = vi.spyOn(Intl, 'Collator')
-    expect(sortAutomationListViewItems(items, null)).toEqual(items)
+    expect(sortAutomationListViewItems(list, null, 'en')).toEqual(list)
     const sort = { field: 'lastRun', direction: 'desc' } as const
-    expect(sortAutomationListViewItems(items, sort)).toEqual(previousOrder(items, sort))
-    expect(sortAutomationListViewItems([], { field: 'name', direction: 'asc' })).toEqual([])
+    expect(sortAutomationListViewItems(list, sort, 'en')).toEqual(previousOrder(list, sort, 'en'))
+    expect(sortAutomationListViewItems([], { field: 'name', direction: 'asc' }, 'en')).toEqual([])
     expect(
-      sortAutomationListViewItems(items.slice(0, 1), { field: 'name', direction: 'asc' })
-    ).toEqual(items.slice(0, 1))
+      sortAutomationListViewItems(list.slice(0, 1), { field: 'name', direction: 'asc' }, 'en')
+    ).toEqual(list.slice(0, 1))
     expect(construct).not.toHaveBeenCalled()
   })
 })
