+++ src/renderer/src/components/automations/automation-list-view.test.ts
@@ -1,7 +1,6 @@
 import { describe, expect, it } from 'vitest'
 import type {
   Automation,
-  AutomationRun,
   AutomationRunStatus,
   ExternalAutomationJob,
   ExternalAutomationManager
@@ -44,31 +43,6 @@ function makeAutomation(overrides: Partial<Automation> = {}): Automation {
     missedRunGraceMinutes: 720,
     createdAt: 1,
     updatedAt: 1,
-    ...overrides
-  }
-}
-
-function makeRun(overrides: Partial<AutomationRun> = {}): AutomationRun {
-  return {
-    id: 'run-1',
-    automationId: 'automation-1',
-    title: 'Zebra job',
-    scheduledFor: 10,
-    status: 'completed',
-    trigger: 'scheduled',
-    workspaceId: 'worktree-1',
-    sessionKind: 'terminal',
-    chatSessionId: null,
-    terminalSessionId: null,
-    terminalPaneKey: null,
-    terminalPtyId: null,
-    outputSnapshot: null,
-    precheckResult: null,
-    usage: null,
-    error: null,
-    startedAt: 20,
-    dispatchedAt: 30,
-    createdAt: 10,
     ...overrides
   }
 }
@@ -118,22 +92,69 @@ function makeExternalEntry(
   }
 }
 
+/** A catalog row with an optional projected last-run status, keyed like a real host row. */
+function makeCatalogRow(
+  id: string,
+  overrides: Partial<Automation> = {},
+  lastRunStatus?: AutomationRunStatus
+): AutomationListRow {
+  return {
+    key: `row|host|${id}`,
+    automation: makeAutomation({ id, ...overrides }),
+    hostLabel: 'This computer',
+    usageSummary: lastRunStatus
+      ? {
+          knownRuns: 1,
+          unavailableRuns: 0,
+          inputTokens: 0,
+          outputTokens: 0,
+          cacheTokens: 0,
+          reasoningOutputTokens: 0,
+          totalTokens: 0,
+          estimatedCostUsd: null,
+          lastRunStatus,
+          lastRunAt: 111
+        }
+      : null
+  }
+}
+
+const rowKey = (id: string): string => `row|host|${id}`
+
 describe('automation-list-view', () => {
   it('counts and detects active filters', () => {
-    expect(isAutomationListFilterActive({ status: 'all', lastRun: 'all', agentIds: [] })).toBe(
-      false
-    )
-    expect(isAutomationListFilterActive({ status: 'paused', lastRun: 'all', agentIds: [] })).toBe(
-      true
-    )
-    expect(countAutomationListFilters({ status: 'paused', lastRun: 'failed', agentIds: [] })).toBe(
-      2
-    )
+    expect(
+      isAutomationListFilterActive({
+        status: 'all',
+        lastRun: 'all',
+        agentIds: []
+      })
+    ).toBe(false)
+    expect(
+      isAutomationListFilterActive({
+        status: 'paused',
+        lastRun: 'all',
+        agentIds: []
+      })
+    ).toBe(true)
+    expect(
+      countAutomationListFilters({
+        status: 'paused',
+        lastRun: 'failed',
+        agentIds: []
+      })
+    ).toBe(2)
   })
 
   it('toggles sort direction and defaults last run to newest first', () => {
-    expect(nextAutomationListSort(null, 'name')).toEqual({ field: 'name', direction: 'asc' })
-    expect(nextAutomationListSort(null, 'lastRun')).toEqual({ field: 'lastRun', direction: 'desc' })
+    expect(nextAutomationListSort(null, 'name')).toEqual({
+      field: 'name',
+      direction: 'asc'
+    })
+    expect(nextAutomationListSort(null, 'lastRun')).toEqual({
+      field: 'lastRun',
+      direction: 'desc'
+    })
     expect(nextAutomationListSort({ field: 'name', direction: 'asc' }, 'name')).toEqual({
       field: 'name',
       direction: 'desc'
@@ -146,82 +167,62 @@ describe('automation-list-view', () => {
 
   it('filters by enabled state and last-run outcome', () => {
     const items = applyAutomationListView({
-      automations: [
-        makeAutomation({ id: 'paused', name: 'Paused', enabled: false }),
-        makeAutomation({ id: 'ok', name: 'Healthy' })
+      rows: [
+        makeCatalogRow('paused', { name: 'Paused', enabled: false }, 'completed'),
+        makeCatalogRow('ok', { name: 'Healthy' }, 'dispatch_failed')
       ],
       externalEntries: [makeExternalEntry()],
-      runs: [
-        makeRun({ automationId: 'paused', status: 'completed' }),
-        makeRun({ automationId: 'ok', status: 'dispatch_failed' })
-      ],
       filter: { status: 'enabled', lastRun: 'failed', agentIds: [] },
-      sort: null
+      sort: null,
+      locale: 'en'
     })
-    expect(items.map((item) => item.id)).toEqual(['ok', 'manager-1:job-1'])
+    expect(items.map((item) => item.id)).toEqual([rowKey('ok'), 'manager-1:job-1'])
   })
 
   it('filters local rows by multiple agents and leaves external rows out of agent scopes', () => {
     const items = applyAutomationListView({
-      automations: [
-        makeAutomation({ id: 'codex-job', agentId: 'codex' }),
-        makeAutomation({ id: 'claude-job', agentId: 'claude' })
+      rows: [
+        makeCatalogRow('codex-job', { agentId: 'codex' }),
+        makeCatalogRow('claude-job', { agentId: 'claude' })
       ],
       externalEntries: [makeExternalEntry()],
-      runs: [],
       filter: { status: 'all', lastRun: 'all', agentIds: ['codex', 'claude'] },
-      sort: null
+      sort: null,
+      locale: 'en'
     })
 
-    expect(items.map((item) => item.id)).toEqual(['codex-job', 'claude-job'])
+    expect(items.map((item) => item.id)).toEqual([rowKey('codex-job'), rowKey('claude-job')])
   })
 
   it('counts an agent filter alongside status and last-run filters', () => {
-    expect(isAutomationListFilterActive({ status: 'all', lastRun: 'all', agentIds: [] })).toBe(
-      false
-    )
     expect(
-      countAutomationListFilters({ status: 'paused', lastRun: 'failed', agentIds: ['codex'] })
+      isAutomationListFilterActive({
+        status: 'all',
+        lastRun: 'all',
+        agentIds: []
+      })
+    ).toBe(false)
+    expect(
+      countAutomationListFilters({
+        status: 'paused',
+        lastRun: 'failed',
+        agentIds: ['codex']
+      })
     ).toBe(3)
   })
 
   it('sorts by name across local and external rows', () => {
     const items = applyAutomationListView({
-      automations: [makeAutomation({ name: 'Zebra job' })],
+      rows: [makeCatalogRow('zebra', { name: 'Zebra job' })],
       externalEntries: [makeExternalEntry({ name: 'Alpha digest' })],
-      runs: [],
       filter: { status: 'all', lastRun: 'all', agentIds: [] },
-      sort: { field: 'name', direction: 'asc' }
+      sort: { field: 'name', direction: 'asc' },
+      locale: 'en'
     })
     expect(items.map((item) => item.name)).toEqual(['Alpha digest', 'Zebra job'])
   })
 
   it('filters catalog rows by status, agent, and the projected last-run status', () => {
-    function makeCatalogRow(
-      id: string,
-      overrides: Partial<Automation>,
-      lastRunStatus?: AutomationRunStatus
-    ): AutomationListRow {
-      return {
-        key: `row|host|${id}`,
-        automation: makeAutomation({ id, ...overrides }),
-        hostLabel: 'This computer',
-        usageSummary: lastRunStatus
-          ? {
-              knownRuns: 1,
-              unavailableRuns: 0,
-              inputTokens: 0,
-              outputTokens: 0,
-              cacheTokens: 0,
-              reasoningOutputTokens: 0,
-              totalTokens: 0,
-              estimatedCostUsd: null,
-              lastRunStatus,
-              lastRunAt: 111
-            }
-          : null
-      }
-    }
     const rows = [
       makeCatalogRow('paused-codex', { enabled: false, agentId: 'codex' }),
       makeCatalogRow('failed-claude', { agentId: 'claude' }, 'dispatch_failed'),
@@ -229,9 +230,10 @@ describe('automation-list-view', () => {
       makeCatalogRow('never-codex', { agentId: 'codex' })
     ]
     const ids = (filter: Partial<AutomationListFilter>) =>
-      filterAutomationListRows(rows, { ...EMPTY_AUTOMATION_LIST_FILTER, ...filter }).map(
-        (row) => row.automation.id
-      )
+      filterAutomationListRows(rows, {
+        ...EMPTY_AUTOMATION_LIST_FILTER,
+        ...filter
+      }).map((row) => row.automation.id)
 
     expect(ids({ status: 'paused' })).toEqual(['paused-codex'])
     expect(ids({ agentIds: ['claude'] })).toEqual(['failed-claude'])
@@ -249,7 +251,10 @@ describe('automation-list-view', () => {
       catalogRef:
         targetId === null
           ? null
-          : { authority: { kind: 'desktop' }, selector: { kind: 'ssh', targetId } },
+          : {
+              authority: { kind: 'desktop' },
+              selector: { kind: 'ssh', targetId }
+            },
       hostLabel: targetId ?? '',
       usageSummary: null
     })
@@ -257,9 +262,10 @@ describe('automation-list-view', () => {
     const keyOf = (row: AutomationListRow): string =>
       row.catalogRef ? hostStableKey(row.catalogRef) : ''
     const ids = (hostStableKeys: readonly string[]) =>
-      filterAutomationListRows(rows, { ...EMPTY_AUTOMATION_LIST_FILTER, hostStableKeys }).map(
-        (row) => row.automation.id
-      )
+      filterAutomationListRows(rows, {
+        ...EMPTY_AUTOMATION_LIST_FILTER,
+        hostStableKeys
+      }).map((row) => row.automation.id)
 
     // Multi-select is any-of; a pre-catalog row names no host and is excluded.
     expect(ids([keyOf(rows[0]), keyOf(rows[1])])).toEqual(['on-a', 'on-b'])
@@ -290,15 +296,22 @@ describe('automation-list-view', () => {
 
   it('sorts by last run newest first and keeps never-run rows last', () => {
     const items = applyAutomationListView({
-      automations: [
-        makeAutomation({ id: 'old', name: 'Old' }),
-        makeAutomation({ id: 'never', name: 'Never' })
+      rows: [
+        makeCatalogRow('old', {
+          name: 'Old',
+          lastRunAt: Date.parse('2026-08-11T09:00:00Z')
+        }),
+        makeCatalogRow('never', { name: 'Never' })
       ],
       externalEntries: [makeExternalEntry({ lastRunAt: '2026-08-12T09:00:00Z' })],
-      runs: [makeRun({ automationId: 'old', dispatchedAt: Date.parse('2026-08-11T09:00:00Z') })],
       filter: { status: 'all', lastRun: 'all', agentIds: [] },
-      sort: { field: 'lastRun', direction: 'desc' }
+      sort: { field: 'lastRun', direction: 'desc' },
+      locale: 'en'
     })
-    expect(items.map((item) => item.id)).toEqual(['manager-1:job-1', 'old', 'never'])
+    expect(items.map((item) => item.id)).toEqual([
+      'manager-1:job-1',
+      rowKey('old'),
+      rowKey('never')
+    ])
   })
 })
