+++ src/renderer/src/components/automations/use-automations-page-list-state.ts
@@ -4,9 +4,12 @@ import { buildExternalAutomationListEntries } from './external-automation-list-e
 import { externalAutomationScopeEntries } from './external-automation-scope-gating'
 import { externalAutomationUncheckedNotice } from './external-automation-unchecked-hosts'
 import {
+  buildAutomationListViewItems,
   filterAutomationListRows,
-  filterExternalAutomationListEntries
+  filterExternalAutomationListEntries,
+  sortAutomationListViewItems
 } from './automation-list-view'
+import { getIntlLocale } from '@/i18n/i18n'
 import { unscopedAutomationListRows } from './automation-list-row-identity'
 import { useAutomationHostCatalog } from './use-automation-host-catalog'
 import { useAutomationListSearch } from './use-automation-list-search'
@@ -28,6 +31,7 @@ export function useAutomationsPageListState({
     failedAuthorityKeys,
     listSearchQuery,
     listFilter,
+    listSort,
     selectedRowKey,
     selectedExternalKey,
     selectedAutomationRuns,
@@ -129,6 +133,21 @@ export function useAutomationsPageListState({
     () => externalAutomationUncheckedNotice(scopedExternal.failures, hostCatalog.entries),
     [hostCatalog.entries, scopedExternal.failures]
   )
+  // Why: a language switch changes collation without touching rows, so the locale
+  // has to reach the memo as a value.
+  const sortLocale = getIntlLocale()
+  const sortedListItems = useMemo(
+    () =>
+      sortAutomationListViewItems(
+        buildAutomationListViewItems({
+          rows: filteredRows,
+          externalEntries: filteredExternalAutomationEntries
+        }),
+        listSort,
+        sortLocale
+      ),
+    [filteredExternalAutomationEntries, filteredRows, listSort, sortLocale]
+  )
 
   return {
     hostCatalog,
@@ -146,6 +165,7 @@ export function useAutomationsPageListState({
     isListSearchQueryTooLarge,
     filteredRows,
     filteredExternalAutomationEntries,
+    sortedListItems,
     hasListItems,
     hasFilteredListItems,
     searchCounts,
